requirements.txt
pymongo==4.6.1
python-dotenv==1.0.1
sentence-transformers==2.5.1
numpy==1.24.3
tiktoken==0.6.0
python-dateutil==2.8.2
scikit-learn==1.3.2
urllib3<2.0.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.2
transformers==4.36.2 